Ketamine Module
Initial Ketamine source, a tool meant to work from Quasar that I made for myself in the first place.. Capable of extracting mods from Quasar's downloads, needs to be put in the same directory as Quasar in order to work properly. Character list is user updatable, and gives info on replaced costumes for each character when the mod is in .../Quasar/Library/Ketamine Module/mods which is the default folder where your mods are extracted in an organized layout, perfect for using ARCadia after putting the folder in your microSD card's /ultimate folder ! Command list is available on the readme.